datamanager/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.7
pydantic-settings>=2.2
fastapi>=0.115
uvicorn>=0.30
python-dotenv>=1.0
